requirements.txt
evoagentx>=0.1.0
chardet>=5.0.0
openai>=1.55.3
dashscope>=1.23.4
fastapi>=0.115.11
uvicorn>=0.34.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
loguru>=0.7.3
pandas>=2.2.3
numpy>=1.26.4
pydantic>=2.9.0
arxiv>=2.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0
websockets>=12.0